+++ tests/unit/enhanced_library/test_enhanced_types.py
@@ -12,27 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from __future__ import absolute_import
-import unittest
+import pytest
 
 from google.cloud.aiplatform.v1beta1.schema.trainingjob import definition
 
 ModelType = definition.AutoMlImageClassificationInputs().ModelType
+test_training_input = definition.AutoMlImageClassificationInputs(
+    multi_label=True,
+    model_type=ModelType.CLOUD,
+    budget_milli_node_hours=8000,
+    disable_early_stopping=False,
+)
 
 
-class EnhancedTypesTests(unittest.TestCase):
-    def setUp(self):
-        self.test_training_input = definition.AutoMlImageClassificationInputs(
-            multi_label=True,
-            model_type=ModelType.CLOUD,
-            budget_milli_node_hours=8000,
-            disable_early_stopping=False,
-        )
+def test_exposes_to_value_method():
+    assert hasattr(test_training_input, "to_value")
 
-    def test_exposes_to_value_method(self):
-        assert hasattr(self.test_training_input, "to_value")
 
-    def test_exposes_from_value_method(self):
-        assert hasattr(self.test_training_input, "from_value")
+def test_exposes_from_value_method():
+    assert hasattr(test_training_input, "from_value")
 
-    def test_exposes_from_map_method(self):
-        assert hasattr(self.test_training_input, "from_map")
+
+def test_exposes_from_map_method():
+    assert hasattr(test_training_input, "from_map")
